% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{as.diaMeasure.matrix}
\alias{as.diaMeasure.matrix}
\title{Coerce matrix to diaMeasure}
\usage{
\method{as.diaMeasure}{matrix}(x, idVars, ...)
}
\arguments{
\item{x}{square matrix}

\item{idVars}{attributes}

\item{...}{parameters passed to other methods}
}
\description{
Coerce matrix to diaMeasure.
}
