% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{as.diaMeasure}
\alias{as.diaMeasure}
\title{Coerce to diaMeasure}
\usage{
as.diaMeasure(x, ...)
}
\arguments{
\item{x}{Object to be coerced}

\item{...}{parameters passed to other methods}
}
\description{
Coertion to diaMeasure, see the corresponding method.
}
