% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{as.matrix.diaMeasure}
\alias{as.matrix.diaMeasure}
\title{Coertion to matrix}
\usage{
\method{as.matrix}{diaMeasure}(x, ...)
}
\arguments{
\item{x}{object of class diaMeasure}

\item{...}{unused}
}
\description{
Coerce the diaMeasure object into a  matrix class. Uses the as.matrix method from the package dist.
}
