% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{diaMeasure}
\alias{diaMeasure}
\title{Compute Dialectometrical measure}
\usage{
diaMeasure(data, formula, value.var, measure = c("lv", "rdi", "ipd",
  "osa", "lv", "dl", "hamming", "lcs"), binary.index = c("jac", "dice",
  "cover"), weight = c(d = 1, i = 1, s = 1, t = 1), q = 1L, p = 0,
  bt = 0, useBytes = FALSE, variable.dist = FALSE)
}
\arguments{
\item{data}{data frame, data.table or object that is coercible by data.table::as.data.table
containing all the variables that are referenced by the \code{formula}. The \code{data} must be in
long format meaning each response must be in a single column. \code{data(dsample)} is an example of the
data format.}

\item{formula}{formula indicating which variables represent identity variables and which measure
variables. See Examples.}

\item{value.var}{character vector of length one indicating which variable of data contains the linguistical
values. See Examples.}

\item{measure}{method: Dissimilarity index, match to ‘"osa"’,
‘"lv"’, ‘"dl"’, ‘"hamming"’, ‘"lcs"’, ‘"rdi"’, ‘"ipd"’.}

\item{binary.index}{binary index to be used with multiple response. For measures ‘"rdi"’ and ‘"ipd"’
"jac" (jaccard) and "dice" (dice) are available. For the rest of measures "cover" (cover set distance)
must be used.}

\item{weight}{pondreation for the distance metrics. For "ipd" weights must be numeric vector of length
one. The higher the weight for "ipd" the less impact very rare linguistical values have. See
Goebsl}

\item{q}{currently unused.}

\item{p}{currently unused.}

\item{bt}{currently unused.}

\item{useBytes}{Perform byte-wise comparison, see
stringdist-encoding (from the stringdist package).}

\item{variable.dist}{logical of length one indicating if the dialectometrical distance needs to be
computed between linguistical identities (FALSE) or linguistical variables (TRUE).}
}
\value{
a vector with dialectometric distances (can be coerced into matrix).
}
\description{
\code{diaMeasure} computes the requested dialectometrical measure between variables of \code{data}.
}
\examples{

data(dsample)

## The linguistical identities are "gender" and "location" meaning that each line
## that belongs to the same location  and gender will be grouped. The distance between
## each gender and each location will be computed
measure <- diaMeasure(dsample, gender + location ~ question, 'answer', 'ird')

## if the linguistical identity is only the gender then all the reponses that each gender has
## given belong to a single group. The distance between genders will be computed.
measure <- diaMeasure(dsample, gender ~ question, 'answer', 'ird')

}
